+++ Systems/System.cs
@@ -12,7 +12,7 @@ public abstract class System
 
     public abstract void Draw();
 
-    public abstract void Update(long delta,InputState input);
+    public abstract void Update(float delta,InputState input);
 
     public abstract void Detach();
 }
